+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 24.1.2
+* Added six new configuration options under the 'limits' interface of 'CountlyConfig':
+* ...
+* ...
+
+* Underlying Android SDK version is 24.1.1
+* Underlying iOS SDK version is 24.1.0
+
 ## 24.1.1
 * Added a new metric for detecting whether or not a device has a hinge for Android